+++ src/com/gnjBook/controller/admin/FileboardAddCtrl.java
@@ -1,5 +1,9 @@
 package com.gnjBook.controller.admin;
 
+import com.gnjBook.dto.Fileboard;
+import com.gnjBook.model.FileboardDAO;
+
+
 import javax.servlet.RequestDispatcher;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -7,13 +11,18 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
+import java.util.List;
 
 @WebServlet("/FileboardAdd.do") // 사용자가 보는 이름
 public class FileboardAddCtrl extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        request.setAttribute("msg", "학습자료를 추가합니다.");
+        String msg = "학습자료를 업로드 합니다.";
+        FileboardDAO dao = new FileboardDAO();
+        List<Fileboard> fileboardList = dao.getFileboardList();
 
+        request.setAttribute("fileboardList", fileboardList);
+        request.setAttribute("msg", msg);
         RequestDispatcher view = request.getRequestDispatcher("/WEB-INF/admin/fileboardAdd.jsp");
         view.forward(request, response);
     }
