+++ web/WEB-INF/admin/adminBoardList.jsp
@@ -21,6 +21,11 @@
                 교재 관리
             </a>
         </li>
+        <li class="nav-item">
+            <a id="admin-file-nav" href="${rootPath}/FileboardListAdmin.do" class="nav-link text-white">
+                학습자료 관리
+            </a>
+        </li>
     </ul>
 </div>
