+++ web/WEB-INF/admin/fileboardList.jsp
@@ -1,0 +1,69 @@
+<%@ page contentType="text/html;charset=UTF-8" pageEncoding="UTF-8" language="java" %>
+<%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core"  %>
+<%@ taglib prefix="fmt" uri="http://java.sun.com/jsp/jstl/fmt" %>
+<%@ page import="java.util.*, java.lang.*" %>
+<%@ page import="java.text.*, java.net.InetAddress" %>
+<c:set var="path" value="<%=request.getContextPath() %>" />
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>학습자료 목록</title>
+    <%@ include file="../../common.jsp"%>
+</head>
+
+<body>
+<%@ include file="../../header.jsp"%>
+<div style="display: flex; min-height: 80vh;">
+    <%@include file="adminBoardList.jsp"%>
+    <div class="container" style="margin-top: 20px;">
+        <h2 class="title">학습자료 목록</h2>
+        <div class="container">
+            <table class="table table-secondary" id="tb1">
+                <thead>
+                <tr>
+                    <th class="item1">번호</th>
+                    <th class="item2">제목</th>
+                    <th class="item3">작성일</th>
+                    <th class="item4">조회수</th>
+                </tr>
+                </thead>
+                <tbody>
+                <c:forEach var="fileboard" items="${fileboardList}" varStatus="status">
+                    <tr>
+                        <td class="item1">${status.count}</td>
+                        <td>
+                            <c:if test="${!empty fileboard.filename1 }">
+                            <img src="${path }/storage/${fileboard.filename1 }" alt="${fileboard.filename1 }" width="100"/>
+                            </c:if>
+                        </td>
+                        <td>
+                            <c:if test="${!empty fileboard.filename2 }">
+                            <img src="${path }/storage/${fileboard.filename2 }" alt="${fileboard.filename2 }" width="100"/>
+                            </c:if>
+                        </td>
+                        <td>
+                            <c:if test="${!empty fileboard.filename3 }">
+                            <img src="${path }/storage/${fileboard.filename3 }" alt="${fileboard.filename3 }" width="100"/>
+                            </c:if>
+                        </td>
+                    </tr>
+                </c:forEach>
+           </tbody>
+       </table>
+    <a class="btn btn-primary" href="${rootPath }/FileboardAdd.do" role="button">자료 추가</a>
+    <nav aria-label="Page navigation example" id="page-nation1">
+        <ul class="pagination">
+            <li class="page-item"><a class="page-link" href="#">Previous</a></li>
+            <li class="page-item"><a class="page-link" href="#">1</a></li>
+            <li class="page-item"><a class="page-link" href="#">2</a></li>
+            <li class="page-item"><a class="page-link" href="#">3</a></li>
+            <li class="page-item"><a class="page-link" href="#">Next</a></li>
+        </ul>
+    </nav>
+    </form>
+</div>
+<%@ include file="../../footer.jsp" %>
+</body>
+</html>
